Added API calls for searching , adding games to user profile and retrieving games for the user

+++ PlayTogether/Shared/DTOs/GamingPlatformDto.cs
@@ -8,7 +8,7 @@ namespace PlayTogether.Shared.DTOs
     {
         public int Id { get; set; }
 
-        public int ApiId { get; set; }
+        public long ApiId { get; set; }
 
         public string Abbreviation { get; set; }
